Bug 415980 - HTML5 attribute validator marks ng-app AngularJS attributes as undefined
A new preference is created for the pattern of attribute names to be ignored by HTMLAttributeValidator
Web->HTML Files->Validation preference page is modified due to allow the ignored attribute names to be shown/edited by users.
New Quick-fix is added due to allow the users to add attribute names and attribute name patterns to the HTML Attribute Validator
ignore list.

Signed-off-by: Victor V Rubezhny <vrubezhny@exadel.com>