Bug 448152 - [validation] Ability to ignore custom html tag (UI part that adds Validation Pref. page changes and new QuickFix)
Now it's able to ignore html element names from validation using comma-separated name patterns, that can be
specified on Validation Preference page, or set by a quick fix.

This is the UI Part that adds the possibility to set the new preference value on Validation Preference page as well as
adds the new QuickFix that allows a user to add new tag names to be ignored in validation.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>