Bug 491928 - Outline view is not refreshed properly when adding objects or array's children
In this fix the notifier method is called when a child has just been
added to an array or object, then I call the notifier method for the
parent pair, to notify that the value of the pair (in this case the
object or array has changed). After making these method calls the
outline view is updated as expected automatically, without reloading the
entire model (as was previously required).

Change-Id: I6803700b42afe8eb5f9882bd794903b9387dbf1f
Signed-off-by: alina <alina@mx1.ibm.com>
Signed-off-by: Victor Rubezhny <vrubezhny@redhat.com>