Bug 536668 - Timeout on startup while loading class "org.eclipse.wst.sse.core.internal.preferences.PreferenceInitializer"
It is in general discouraged to do lot of work in start() method. In
particular case we can have a deadlock because we are trying to acquire
lock on TaskScanningJob, while holding an activation lock in OSGI. If
another thread manages to lock TaskScanningJob first and will try to
activate our bundle to init bundle preferences, we deadlock.

Solution is simply to run TaskScanningScheduler.startup() in a
background job that does not run within this bundle activation sequence.

Change-Id: Id8b531da95b974c77e598cf65886f44647af1dae
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>